Add module document_status to simplify code for less redundancy
