+++ part-3/damage_report.rb
@@ -1,21 +1,11 @@
+require_relative "document_status"
+
 class DamageReport
-  attr_reader :description
-  attr_accessor :assigned_employee
+  include DocumentStatus
 
   def initialize(args = {})
     @description = args.fetch(:description) { "Unknown damage" }
     @resolved = false
   end
 
-  def resolved?
-    !!@resolved
-  end
-
-  def close
-    @resolved = true
-  end
-
-  def assigned?
-    !!@assigned_employee
-  end
 end
