+++ part-3/incident_report.rb
@@ -1,6 +1,9 @@
+require_relative "document_status"
+
 class IncidentReport
-  attr_reader :description, :reporter
-  attr_accessor :assigned_employee
+  include DocumentStatus
+
+  attr_reader :reporter
 
   def initialize(args = {})
     @description = args.fetch(:description) { "Unknown damage" }
@@ -8,16 +11,5 @@ class IncidentReport
     @resolved = false
   end
 
-  def resolved?
-    !!@resolved
-  end
-
-  def close
-    @resolved = true
-  end
-
-  def assigned?
-    !!@assigned_employee
-  end
 end
